Action: addition of reflection on the unstructured grid and upgrade of TEST441 to TEST451, and bugfix for NLPROP default value in ST2
Type: enhancement and bugfix                           

Ticket: #53 

Milestone: 

Command: svn commit 

Author: Fabrice Ardhuin

Details: 
Updates here include 
1) a new treatment of the boundary conditions in W3PROFSMD based on the
definition of shoreline directions in w3triamd.ftn, which is updated
when the water level is updated, and reflection coefficients computed 
in w3ref1md.ftn when the switch REF1 is used. The basic algorithm is 
very simple: this is a Dirichlet boundary condition. Tests and results
are shown in a paper submitted to JGR. A preprint of that can be found
here: 
http://wwz.ifremer.fr/iowaga/content/download/54671/772821/file/Ardhuin_Roland_reflection_submitted_2col.pdf
is is under revision. 
This change impacts w3profsmd.ftn, w3triamd.ftn, ww3_grid.ftn,
w3gdatmd.ftn, and w3ref1md.ftn. 

The update to w3profsmd.ftn will follow in the next commit. 

2) an upgrade of the swell dissipation parameterization in w3src4md.ftn
, with the addition of a TANH(RE/SWELLF(7)) smoother from laminar to
turbulent. 

3) a bug fix for the NLPROP default value to be used in ST2. 



Former-commit-id: 4b43779fb65d3406e9266f5466a2190ef41b0a03