Reintegrating branches/bugfix_3dout_points which corrects bugs with 3D output.
This was first addressed in v79534, but this introduced other bugs with multi using point and partitioning processors.  So these changes were walked back (these bugs appeared in wmm3_test_03 with gridests d2 and e).  

The original problem appears to be that we were using E3DF, P2MSF and US3DF (flag variables to allocate 3D output arrays which are set in mod_def files) in W3READFLGD and W3FLGRDFLAG, which had not yet been set, since mod_def files had yet to be read. 

To address the original issue, IF statements involving E3DF, P2MSF and US3DF in W3READFLGD and W3FLGRDFLAG were removed.  This should be okay because W3FLGRDUPDT will be called in W3INIT after the mod_def grid files are read to update the output flags based on E3DF, etc.  



Former-commit-id: 8f83351bb65075773ff914fea8781f9b58f285ae