Fix for w3servmd:EXTCDE.
Replace the use of MPI_COMM_WORLD in MPI_BARRIER with a communicator
that is passed in as an optional argument.  The use of MPI_COMM_WORLD
in the MPI_BARRIER is potential for a dead-lock, especially in a coupled
system, where not all processes invoke EXTCDE.

Notes:
This commit is a merge of r855 from NRL/WW3/branches/esmf2_model.



Former-commit-id: a267f955a11c8ff0c9504fe930d72f7609849edf