Bug fix for ww3_ounf.ftn (this fixes ticket #226).
The NCEP reintegration of branches/ustokes to the trunk introduced
a compile-time error in ww3_ounf.ftn.  The USSP_WN variable is defined
as use associated from the W3ADATMD module.  The USSP_WN variable is
defined in the W3GDATMD module, not the W3ADATMD module.  Since the
W3GDATMD module is used without ONLY, the fix is to remove USSP_WN
from the USE W3GDATMD block.



Former-commit-id: ec5f4cbed1bae97f9f3cc17ddc58ed29e1806a7a