Branch: ww3/esmf2 Action: Bug fix for W3IORS with MPI -- segfault during restart read issue Ticket: #232
Description of the error:
The use of MPI_COMM_WAVE as the MPI communicator in the MPI_SCATTER operation
in W3IORSMD:W3IORS is incorrect in that MPI_COMM_WAVE includes all MPI processes
(NTPROC -- computational and I/O) but the send buffer (VGBUFF) being scattered
is only allocated using the number of computational processes (NAPROC).  This error
causes the MPI_SCATTER operation to exceed the upper memory bound of the send
buffer, which sometimes results in a program crash with SEGFAULT during the
restart read.  Because the SEGFAULT occurs in the MPI_SCATTER, no useful traceback
information is obtained.

Description of the fix:
The fix is to use MPI_COMM_WCMP as the MPI communicator in the MPI_SCATTER call.
The MPI_COMM_WCMP communicator only includes the computational MPI processes (NAPROC).
Since MPI_COMM_WCMP does not include the I/O processes, a conditional is added around
the read and scatter block so that only computational processes participate.



Former-commit-id: 6948a45aad91058479ceec868b777759f69ba1d0