Bugs that affect messages about parsing the input files.
There are three bug fixes (A, B, C) for ww3_grid and ww3_shel in two files ( ww3_grid.ftn and w3iogomd.ftn) .

A) A temporary file ww3_grid.scratch is opened not only for old-stype input file ww3_grid.inp, but also when input in ww3_grid.nml is present. The bugfix does not open this file in the latter case.

B) In the standard output from ww3_grid  info is missing about the last outbound file for one-way nesting. The solution is to let NPO=0 mark that info has been read about the last file, and use this value to exit the file message loop.

C) When ww3_shel is running under MPI and and encounters an error in the list of fields specified in ww3_shel.inp or ww3_shel.nml for the output to out_grd.ww3, it attempts writing a WARNING to a file id NDSE == -1. This causes the program to abort on some systems. @MicahDavid was the first to debug this. A file id NDSE /= -1 is defined only for the parallel process to write output ( IAPROC .EQ. NAPOUT ). The solution is to use a condition "IF (NDSEN .GE. 0 " throughout both the subroutines W3READFLGRD and W3FLGRDFLAG. This condition already exists in the subroutines NEXTLN and W3FLDO. This bugfix will probably also apply for ww3_multi.