Add ability to build WW3 as UFS submodule
This required a couple of changes with relative paths changes (CURRENT_SOURCE_DIR vs. CMAKE_SOURCE_DIR) when UFS CMakeLists is used.

Because UFS already defines the 'esmf' target a check was added to skip the search for ESMF if the target already exists.

And include the WW3 Fortran modules in the WW3::WW3 target with target_include_directories.