compile and grd.out fixes
* fix github action compile failure
* create new switch setting specifically to allow uwm netcdf
output, otherwise native binary output will be written. This is not
as flexible as initially implmented, which allowed netcdf gridded
output as a config flag. We reverted that ability in w3iogoncmd to
debug the perf issues on cesm side. UWM netcdf griddded output can
be turned on with compile-time switch UWMNCOUT
* all baselines pass run against shel_wavebl_20220214